+++ packages/ranger-ui/src/CookieConsent/index.ts
@@ -1,0 +1,2 @@
+export type { CookieConsentProps } from './cookieConsent'
+export { default as CookieConsent } from './cookieConsent'