+++ packages/ranger-ui/src/ErrorBoundary/index.ts
@@ -1,0 +1,2 @@
+export type { ErrorBoundaryProps } from './errorBoundary'
+export { default as ErrorBoundary } from './errorBoundary'