+++ packages/ranger-ui/src/IntersectionScroll/index.ts
@@ -1,0 +1,2 @@
+export type { IntersectionScrollProps } from './intersectionScroll'
+export { default as IntersectionScroll } from './intersectionScroll'